Refactor to make tabs
